refactor: Audience, Bag 클레스 병합
