+++ src/main/java/com/wanted/preonboarding/theater/service/handler/Audience.java
@@ -1,11 +1,43 @@
 package com.wanted.preonboarding.theater.service.handler;
 
-public class Audience {
-    private final Bag bag;
+import org.hibernate.mapping.Bag;
 
-    public Audience(Bag bag){
-        this.bag = bag;
+import com.wanted.preonboarding.theater.service.model.RequestMessage;
+
+public class Audience {
+    private Long money;
+    private final Invitation invitation;
+    private Ticket ticket;
+
+    public Audience(long money){
+        this(null, money);
+    }
+    public Audience(Invitation invitation, long money){
+        this.invitation = invitation;
+        this.money = money;
     }
 
-    public Bag getBag(){ return bag;}
+    public static Audience of(RequestMessage requestMessage){
+        Invitation invitation = null;
+        if (requestMessage.getHasInvitation()) {
+            invitation = new Invitation();
+        }
+        return new Audience(invitation, requestMessage.getMoney());
+    }
+
+    public boolean hasInvitation() {
+        return invitation != null;
+    }
+    public boolean hasTicket() {
+        return ticket != null;
+    }
+    public void receiveTicket(Ticket ticket) {
+        this.ticket = ticket;
+    }
+    public void payFee(long money) {
+        this.money -= money;
+    }
+    public void refundFee(long money) {
+        this.money += money;
+    }
 }
