--- src/main/java/com/wanted/preonboarding/theater/service/handler/Bag.java
@@ -1,31 +1,0 @@
-package com.wanted.preonboarding.theater.service.handler;
-
-public class Bag {
-    private Long amount;
-    private final Invitation invitation;
-    private Ticket ticket;
-
-    public Bag(long amount){
-        this(null, amount);
-    }
-    public Bag(Invitation invitation, long amount){
-        this.invitation = invitation;
-        this.amount = amount;
-    }
-
-    public boolean hasInvitation() {
-        return invitation != null;
-    }
-    public boolean hasTicket() {
-        return ticket != null;
-    }
-    public void setTicket(Ticket ticket) {
-        this.ticket = ticket;
-    }
-    public void minusAmount(long amount) {
-        this.amount -= amount;
-    }
-    public void plusAmount(long amount) {
-        this.amount += amount;
-    }
-}
